pyproject.toml
[tool.poetry]
name = "mailboat"
version = "0.0.1-alpha.0"
description = "a email server solution for organisation"
authors = ["The Mailboat Contributors"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
aiosmtpd = "^1.4.2"
pymap = "^0.23.1"
tornado = "^6.1"
unqlite = "^0.9.2"
pyspf = "^2.0.14"
dkimpy = "^1.0.5"
flanker = "^0.9.11"
aiosmtplib = "^1.1.4"
pynacl = "^1.4.0"
httpx = "^0.18.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.812"
pytest = "^6.2.3"
pytest-asyncio = "^0.14.0"
pytest-profiling = "^1.7.0"
pytest-cov = "^2.11.1"
aioimaplib = "^0.9.0"
pre-commit = "^2.12.1"
pdoc3 = "^0.9.2"

[tool.black]
line-length = 88
include = '\.pyi?$'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
